(KDTree) convert to vector of pointers
List of Nodes was used so that list could be reused, so that memory would not be allocated again each resampling. But a node contains pointers to elements in the list, so on resize, these pointers became invalid. By having a list of pointers, while reusing the memory, when already allocated, the same performance is kept, while being able to resize the list